Refactor B1 tab
